+++ src/components/Router/StepsRouter.js
@@ -7,6 +7,17 @@ import theme from '../Theme/style.css'
 import { withFullScreenState } from '../FullScreen'
 
 class StepsRouter extends Component {
+  componentDidUpdate(prevProps) {
+    // Re-focus to container is needed for accessibility to have a correct
+    // tabbing order, and should be triggered when...
+    if (this.container &&
+        // ...step changes (for tabbing order to start from the top)
+        (prevProps.step !== this.props.step) ||
+        // ..."full screen" mode changes (e.g. for enlarged image preview)
+        (prevProps.isFullScreen !== this.props.isFullScreen)) {
+      this.container.focus()
+    }
+  }
 
   trackScreen = (screenNameHierarchy, properties = {}) => {
     const { step } = this.currentComponent()
@@ -25,12 +36,24 @@ class StepsRouter extends Component {
 
     return (
       //TODO: Wrap CurrentComponent in themeWrap HOC
-      <div className={classNames(theme.step,{[theme.fullScreenStep]: isFullScreen})}>
-        <NavigationBar back={back} disabled={disableNavigation} className={theme.navigationBar}/>
-        <div className={classNames(theme.content,{
-          [theme.fullScreenContentWrapper]: isFullScreen,
-          [theme.scrollableContent]: !isFullScreen
-        })}>
+      <div
+        ref={node => this.container = node}
+        tabIndex={-1}
+        className={classNames(theme.step, {
+          [theme.fullScreenStep]: isFullScreen
+        })}
+      >
+      <NavigationBar
+        back={back}
+        disabled={disableNavigation}
+        className={theme.navigationBar}
+      />
+        <div
+          className={classNames(theme.content, {
+            [theme.fullScreenContentWrapper]: isFullScreen,
+            [theme.scrollableContent]: !isFullScreen
+          })}
+        >
           <CurrentComponent {...{...options, ...globalUserOptions, ...otherProps, back}}
             trackScreen={this.trackScreen} />
         </div>