+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 ### Added
 - UI: Accessibility - Make H1 readable by screen readers
 - UI: Accessibility - Make buttons/links readable by screen readers, allow tabbing to them
+- UI: Accessibility - Sort out order of items when tabbing through the content of each step
 - UI: Accessibility - Announce page transition when screen changes
 
 ### Changed